Moving the executable and cleaning up for V1
Removing references to PNGQuant. deleted the executable.js file as it
makes more since to have that information defined in the package.json
file. Added comments to package.json. Added a small library to allow
parsing of json with comments. Updated about modal. Still need to make
it's content dynamic and the modal functional.